requirements.txt
google-generativeai
rich
python-dotenv
typer
pytest
